% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/connect_database.R
\name{connect_database}
\alias{connect_database}
\title{connection_database}
\usage{
connect_database(workshop_dir)
}
\arguments{
\item{workshop_dir}{Directory dir of workshop}
}
\value{
Database connection object
}
\description{
connection_database
}

