% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_database.R
\name{create_database}
\alias{create_database}
\title{create_database}
\usage{
create_database(workshop_dir)
}
\arguments{
\item{workshop_dir}{Directory dir of workshop}
}
\description{
create database for user according to configure file.
if only the user have database permission to create database.
}

