% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_cpu_cores.R
\name{get_cpu_cores}
\alias{get_cpu_cores}
\title{get_cpu_cores}
\usage{
get_cpu_cores(workshop_dir)
}
\arguments{
\item{workshop_dir}{Directory of workshop}
}
\value{
CPU cores that can be used
}
\description{
get_cpu_cores
}

