% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/import_sql.R
\name{import_sql}
\alias{import_sql}
\title{import_sql}
\usage{
import_sql(workshop, sql_file)
}
\arguments{
\item{workshop}{Directory of workshop}

\item{sql_file}{The path of SQL file}
}
\value{
Boolean
}
\description{
Import SQL into database with a given SQL file.
}

