% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_configure_file.R
\name{load_configure_file}
\alias{load_configure_file}
\title{load_configure_file}
\usage{
load_configure_file(workshop_dir)
}
\arguments{
\item{workshop_dir}{Dir of workshop}
}
\value{
A list of configure
}
\description{
Load configure file into list
}

