% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_database_configure.R
\name{load_database_configure}
\alias{load_database_configure}
\title{load_database_configure}
\usage{
load_database_configure(workshop_dir)
}
\arguments{
\item{workshop_dir}{dir of workshop}
}
\value{
a list of configure
}
\description{
load_database_configure
}

