% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_species_setting.R
\name{load_species_setting}
\alias{load_species_setting}
\title{load_species_configure}
\usage{
load_species_setting(workshop_dir, species_name)
}
\arguments{
\item{workshop_dir}{Dir of workshop}

\item{species_name}{Name of species which user want to load configure}
}
\value{
a list of configure
}
\description{
load_species_configure
}

