\name{sdmengine.common-package}
\alias{sdmengine.common-package}
\alias{sdmengine.common}
\docType{package}
\title{
sdmengine.common
}
\description{
Provide common function for SDM-serial software
}
\details{
\tabular{ll}{
Package: \tab sdmengine.common\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-01-13\cr
License: \tab AGPL\cr
}
load_configure_file, load_database_configure, load_species_setting
}
\author{
Xiaoquan Kong, Renyan Duan, Minyi Huang

Maintainer: Xiaoquan Kong <u1mail2me@gmail.com>
}

\keyword{ package }

